+++ web/compile_wasm.sh
@@ -19,8 +19,9 @@ em++ \
 ../../src/filters/filters.cpp \
 ../../src/filters/autogain.cpp \
 ../../src/filters/echo_cancellation.cpp \
--O3 \
+-s MODULARIZE=1 -s EXPORT_NAME="'RecorderModule'" \
 -msimd128 -msse3 \
+-O3 \
 -s "EXPORTED_RUNTIME_METHODS=['ccall','cwrap']" \
 -s "EXPORTED_FUNCTIONS=['_free', '_malloc']" \
 -s EXPORT_ALL=1 -s NO_EXIT_RUNTIME=1 \