+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+## 0.9.3
+- added `autoGain` filter
+- breaking change: 
+  - the `init` method is now async
+
 ## 0.9.2
-* fix: the stop was instead einit the device
-* fix: removed dialog when stopping stream in the example
+- fix: the stop was instead einit the device
+- fix: removed dialog when stopping stream in the example
 
 ## 0.9.1
-* breaking change: 
+- breaking change: 
   - renamed `startListen` to `start`
   - renamed `stopListen` to `stop`
   - renamed `isDeviceStartedListen` to `isDeviceStarted`